Fix iterator node resize stuff causing re-renders on drag, fix drag'n'drop position when zoomed
