+++ src/components/node/InteratorNode.jsx
@@ -1,9 +1,8 @@
 /* eslint-disable react/prop-types */
 /* eslint-disable import/extensions */
 import {
-  Box, Center, Text, useColorModeValue, VStack,
+  Center, Text, useColorModeValue, VStack,
 } from '@chakra-ui/react';
-import { Resizable } from 're-resizable';
 import React, {
   memo, useContext, useEffect, useMemo, useState,
 } from 'react';
@@ -11,42 +10,11 @@ import checkNodeValidity from '../../helpers/checkNodeValidity.js';
 import getAccentColor from '../../helpers/getNodeAccentColors.js';
 import { GlobalContext } from '../../helpers/GlobalNodeState.jsx';
 import shadeColor from '../../helpers/shadeColor.js';
+import IteratorNodeBody from './IteratorNodeBody.jsx';
 import NodeFooter from './NodeFooter.jsx';
 import NodeHeader from './NodeHeader.jsx';
 import NodeInputs from './NodeInputs.jsx';
 import NodeOutputs from './NodeOutputs.jsx';
-
-const createGridDotsPath = (size, fill) => <circle cx={size} cy={size} r={size} fill={fill} />;
-
-const DotPattern = ({ id }) => {
-  const gap = 15;
-  const size = 0.5;
-  const scaledGap = gap * 1;
-  const path = createGridDotsPath(size, '#81818a');
-  const patternId = `pattern-${id}`;
-
-  return (
-    <svg
-      style={{
-        width: '100%',
-        height: '100%',
-        borderRadius: '0.5rem',
-      }}
-    >
-      <pattern
-        id={patternId}
-        x={6}
-        y={6}
-        width={scaledGap}
-        height={scaledGap}
-        patternUnits="userSpaceOnUse"
-      >
-        {path}
-      </pattern>
-      <rect x="0" y="0" width="100%" height="100%" fill={`url(#${patternId})`} />
-    </svg>
-  );
-};
 
 const blankSchema = {
   inputs: [],
@@ -69,14 +37,12 @@ const getSchema = (availableNodes, category, type) => {
 
 const IteratorNode = ({ data, selected }) => {
   const {
-    edges, availableNodes, zoom, useIteratorSize,
+    edges, availableNodes,
   } = useContext(GlobalContext);
 
   const {
-    id, inputData, isLocked, category, type,
+    id, inputData, isLocked, category, type, iteratorSize,
   } = useMemo(() => data, [data]);
-
-  const [{ width, height }, setIteratorSize, defaultSize] = useIteratorSize(id);
 
   // We get inputs and outputs this way in case something changes with them in the future
   // This way, we have to do less in the migration file
@@ -106,16 +72,6 @@ const IteratorNode = ({ data, selected }) => {
     }
   }, [inputData, edges.length]);
 
-  // eslint-disable-next-line no-unused-vars
-  const [showMenu, setShowMenu] = useState(false);
-  // const [menuPosition, setMenuPosition] = useState({});
-
-  // useEffect(() => {
-  //   if (!selected) {
-  //     setShowMenu(false);
-  //   }
-  // }, [selected]);
-
   return (
     <>
       <Center
@@ -126,9 +82,6 @@ const IteratorNode = ({ data, selected }) => {
         py={2}
         boxShadow="lg"
         transition="0.15s ease-in-out"
-        onClick={() => {
-          // setShowMenu(false);
-        }}
       >
         <VStack minWidth="240px">
           <NodeHeader
@@ -151,62 +104,10 @@ const IteratorNode = ({ data, selected }) => {
               ITERATION
             </Text>
           </Center>
-          <Resizable
-            className="nodrag"
-            defaultSize={defaultSize}
-            minWidth="280px"
-            minHeight="280px"
-            draggable={false}
-            enable={{
-              top: false,
-              right: true,
-              bottom: true,
-              left: false,
-              topRight: false,
-              bottomRight: true,
-              bottomLeft: false,
-              topLeft: false,
-            }}
-            scale={zoom}
-            style={{
-              margin: 8,
-              marginBottom: 0,
-            }}
-            size={{ width, height }}
-            onResizeStop={(e, direction, ref, d) => {
-              console.log('resize stop called');
-              setIteratorSize({
-                width: width + d.width,
-                height: height + d.height,
-              });
-            }}
-          >
-            <Box
-              className="nodrag"
-              draggable={false}
-              // bg={useColorModeValue('gray.200', 'gray.800')}
-              // p={2}
-              h="full"
-              w="full"
-              my={0}
-              // boxShadow="inset 0 0 15px var(--chakra-colors-gray-700)"
-              // borderWidth={4}
-              borderColor="gray.700"
-              // borderRadius="xl"
-            >
-              <Box
-                bg={useColorModeValue('gray.200', 'gray.800')}
-                h="full"
-                w="full"
-                borderWidth={1}
-                borderColor="gray.600"
-                borderRadius="lg"
-              >
-                <DotPattern id={id} />
-              </Box>
-              {/* Test */}
-            </Box>
-          </Resizable>
+          <IteratorNodeBody
+            id={id}
+            iteratorSize={iteratorSize}
+          />
           {outputs.length > 0 && (
           <Center>
             <Text fontSize="xs" p={0} m={0} mt={-1} mb={-1} pt={-1} pb={-1}>
